requirements.txt
fastapi
uvicorn
pandas
numpy
python-dotenv
supabase
torch
scikit-learn
